Provide access to the current state when extracting the 'id' needed
In non-trivial cases the id of the entity isn't available in the action itself, but rather needs
to be looked up through another path. If that path needs to go to redux, we need to have access
to the _same_ state that we're modifying, and cannot necessarily just assume we can access the
state somehow again.

Provide the `state` to the `GetId` function as a second parameter.