Try to return the unmodified state object when the item reducer hasn't done anything
Essentially the `getId` function determines whether the item reducer is executed or not. Depending
on how the actual actions look like it is possible that the reducer in fact will not recognize the action
at all, and just return the provided state. In that case we would still return a new state object, which
in turn adds load on the GC mechanisms of node.

This commit adds a very simple identity check, and if the reducer has returned the same state instance
as we provided to it, then we return the original state as well rather than building up a new state object.